[Patrick] - #428: introduces custom_format_to_moment helper
Hexo does not have a unified way to deal with dates.
Sometimes dates come as strings from the front matter file, sometimes as
Date objects.
Sometimes it considers the timezone set in _config.yml, sometimes it
considers a completely crazy timezone.
Because of that from now on (and while hexo does not find its way) we
are going to have '####' at the end of dates in our front matter files
so that hexo will not even try to parse the date.
It is responsibility of custom_format_to_moment helper to parse the
dates with '####' and return a moment object with the timezone set in
_config.yaml