[Patrick, Wandy] FoZAz4Rp uses hexo's partial helper instead of mixins
We are facing issues in productions that seem related to the use of
mixins defined in files different from the file in which the layout
is defined. Because of that we are replacing the use of jade's mixins
by hexo's partials in the hope that this will fix the issue.