[Patrick] 27E3rYvL adds link to the newsletter subscription form
This approach is feature toggled since we still want to have
the form embedded in the sidebar without using an iframe (because of the
different style of allin's form).

We are going with the link approach while there is no support to HTTPS
and CORS headers allowing our domain at allin's side.